Changed NN in-memory representation from fp32 to fp16

+++ src/OpenCL.h
@@ -114,7 +114,7 @@ public:
         return m_layers.size();
     }
 
-    void forward(const std::vector<float>& input, std::vector<float>& output);
+    void forward(const std::vector<half>& input, std::vector<half>& output);
 
 private:
     void push_weights(size_t layer, const std::vector<float> & weights) {
